+++ OverflowJOLJAK/Source/OverflowJOLJAK/EnemyCharacter.h
@@ -24,17 +24,17 @@ public:
     int32 EnemyId = -1;
 
     UPROPERTY(BlueprintReadOnly, Category = "Enemy")
-    int32 MaxHp = 100;
+    float MaxHp = 100.0f;
 
     UPROPERTY(BlueprintReadOnly, Category = "Enemy")
-    int32 CurrentHp = 100;
+    float CurrentHp = 100.0f;
 
     // 서버로부터 새 HP를 받았을 때 호출됨 (NetSyncComponent가 호출).
     // 기본 구현은 CurrentHp만 갱신. 블루프린트에서 이 이벤트를 오버라이드해서
     // 피격 이펙트/히트 리액션 애니메이션을 재생하면 된다.
     UFUNCTION(BlueprintNativeEvent, Category = "Enemy")
-    void OnHpChanged(int32 NewHp, int32 PreviousHp);
-    virtual void OnHpChanged_Implementation(int32 NewHp, int32 PreviousHp);
+    void OnHpChanged(float NewHp, float PreviousHp);
+    virtual void OnHpChanged_Implementation(float NewHp, float PreviousHp);
 
     // 서버가 이 몬스터의 HP가 0 이하가 되어 제거했다고 알려왔을 때 호출됨.
     // 기본 구현은 즉시 Destroy(). 블루프린트에서 오버라이드해서 사망 애니메이션/파티클을
